+++ build.properties
@@ -1,4 +1,4 @@
 minecraft.version=1.8.9
-forge.version=11.15.0.1684
+forge.version=11.15.0.1692
 
 srails.version = 0.0.1
