--- src/main/resources/assets/srails/models/item/emitter.json
@@ -1,10 +1,0 @@
-{
-  "parent":"srails:block/emitter",
-  "display": {
-    "thirdperson": {
-      "rotation": [ 10, -45, 170 ],
-      "translation": [ 0, 1.5, -2.75 ],
-      "scale": [ 0.375, 0.375, 0.375 ]
-    }
-  }
-}
